Refactoring: extract one single encoder which contains edit and code before encoders inside.
