+++ neural_editor/seq2seq/EncoderDecoder.py
@@ -1,10 +1,8 @@
 from typing import Tuple
 
-import torch
-from torch import nn
 from torch import Tensor
+from torch import nn
 
-from edit_representation.sequence_encoding import EditEncoder
 from neural_editor.seq2seq import Generator, Batch
 from neural_editor.seq2seq.decoder import Decoder
 from neural_editor.seq2seq.encoder import Encoder
@@ -15,13 +13,11 @@ class EncoderDecoder(nn.Module):
     A standard Encoder-Decoder architecture. Base for this and many other models.
     """
 
-    def __init__(self, encoder: Encoder, decoder: Decoder, edit_encoder: EditEncoder,
+    def __init__(self, encoder: Encoder, decoder: Decoder,
                  target_embed: nn.Embedding, generator: Generator) -> None:
         super(EncoderDecoder, self).__init__()
-        self.edit_final = None
         self.encoder = encoder
         self.decoder = decoder
-        self.edit_encoder = edit_encoder
         self.target_embed = target_embed
         self.generator = generator
 
@@ -37,34 +33,11 @@ class EncoderDecoder(nn.Module):
                  [B, TrgSeqLen, DecoderH]
         ]
         """
-        edit_final, encoder_output, encoder_final = self.encode(batch)
+        edit_final, encoder_output, encoder_final = self.encoder(batch)
         decoded = self.decode(batch, edit_final, encoder_output,
                               encoder_final, batch.src_mask,
                               batch.trg, batch.trg_mask, None)
         return decoded
-
-    def set_edit_representation(self, sample: Batch) -> None:
-        """
-        Fixates edit_final vector. Used for one-shot learning.
-        :param sample: sample from which construct edit representation, it is batch with size 1
-        :return: nothing
-        """
-        self.edit_final = self.encode_edit(sample)
-
-    def unset_edit_representation(self) -> None:
-        """
-        Unset edit representation. Turns off one-shot learning mode.
-        :return: nothing
-        """
-        self.edit_final = None
-
-    def encode_edit(self, batch: Batch) -> Tuple[Tensor, Tensor]:
-        """
-        Returns edit representations (edit_final) of samples in the batch.
-        :param batch: batch to encode
-        :return: Tuple[[NumLayers, B, NumDirections * DiffEncoderH], [NumLayers, B, NumDirections * DiffEncoderH]]
-        """
-        return self.edit_encoder.encode_edit(batch)
 
     def encode(self, batch: Batch) -> Tuple[Tuple[Tensor, Tensor], Tensor, Tuple[Tensor, Tensor]]:
         """
@@ -76,12 +49,7 @@ class EncoderDecoder(nn.Module):
             Tuple[[NumLayers, B, NumDirections * SrcEncoderH], [NumLayers, B, NumDirections * SrcEncoderH]]
         ]
         """
-        if self.edit_final is None:
-            edit_final = self.encode_edit(batch)
-        else:
-            edit_final = self.edit_final
-        encoder_output, encoder_final = self.encoder(batch)
-        return edit_final, encoder_output, encoder_final
+        return self.encoder(batch)
 
     def decode(self, batch: Batch, edit_final: Tuple[Tensor, Tensor],
                encoder_output: Tensor, encoder_final: Tuple[Tensor, Tensor],
